# api/openapi.yaml
openapi: 3.0.3
info:
  title: SignalWatch API
  version: 2.5.0-credentials
  description: |
    SignalWatch M4 HTTP API contract. All timestamps are UTC RFC3339 values.
    The public API covers identity, profile settings, the controlled source
    catalog, subscription CRUD, and the current user's matched arXiv papers.
    Background local matching provides explainable match reasons. M4 adds the
    subscription-scoped daily Digest (scheduled in the user timezone) and email delivery Worker; mail delivery operations
    are intentionally not exposed as public endpoints. User-supplied provider credentials provide optional paper summaries and cached Digest enrichment; Bounded Agents provide persistent subscription drafting and evidence-based paper conversations. Read-only operations endpoints expose safe runtime
    status to database-authorized operator accounts; no administrative mutation
    or log-download endpoint is exposed.
...
security:
  - bearerAuth: []
paths:
  /api/v2/ai/credentials:
    get:
      operationId: listAICredentials
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AIConfiguration'
        default:
          $ref: '#/components/responses/AIError'
    post:
      operationId: createAICredential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, provider, model, api_key]
              properties:
                name: {type: string, minLength: 1, maxLength: 80}
                provider: {type: string}
                model: {type: string}
                api_key: {type: string, writeOnly: true, minLength: 8, maxLength: 1024}
      responses:
        '201':
          description: Verified credential created with a new identity; existing keys are preserved.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AIConfiguration'}
        default:
          $ref: '#/components/responses/AIError'
  /api/v2/ai/credentials/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
    get:
      operationId: getAICredential
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIConfiguration'
        default:
          $ref: '#/components/responses/AIError'
    put:
      operationId: putAICredential
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIConfiguration'
        default:
          $ref: '#/components/responses/AIError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - name
              - model
              properties:
                name: {type: string, minLength: 1, maxLength: 80}
                model:
                  type: string
                api_key:
                  type: string
                  writeOnly: true
                  maxLength: 1024
      parameters:
      - name: If-Match
        in: header
        schema:
          type: string
        required: true
        description: Credential generation:version in double quotes.
    delete:
      operationId: deleteAICredential
      responses:
        '204':
          description: Completed; no response body.
        default:
          $ref: '#/components/responses/AIError'
      parameters:
      - name: If-Match
        in: header
        required: true
        schema:
          type: string
  /api/v2/ai/credentials/{id}/test:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
    post:
      operationId: testAICredential
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIConfiguration'
        default:
          $ref: '#/components/responses/AIError'
  /api/v2/ai/default-selection:
    get:
      operationId: getAIDefaultSelection
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIConfiguration'
        default:
          $ref: '#/components/responses/AIError'
    put:
      operationId: putAIDefaultSelection
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIConfiguration'
        default:
          $ref: '#/components/responses/AIError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - provider
              - generation
              - version
              properties:
                provider:
                  type: string
                model:
                  type: string
                generation:
                  type: string
                version:
                  type: integer
                  minimum: 1
  /api/v2/agent/conversations:
    get:
      operationId: listAgentConversations
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - items
                - page
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Conversation'
                  page:
                    type: integer
        default:
          $ref: '#/components/responses/AIError'
      parameters:
      - name: kind
        in: query
        schema:
          type: string
          enum:
          - subscription
          - paper
      - name: paper_id
        in: query
        schema:
          type: integer
          minimum: 1
      - name: page
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100000
    post:
      operationId: createAgentConversation
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        default:
          $ref: '#/components/responses/AIError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - kind
              properties:
                kind:
                  type: string
                  enum:
                  - subscription
                  - paper
                paper_id:
                  type: integer
                  minimum: 1
  /api/v2/agent/conversations/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
    get:
      operationId: getAgentConversation
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        default:
          $ref: '#/components/responses/AIError'
    delete:
      operationId: deleteAgentConversation
      responses:
        '204':
          description: Completed; no response body.
        default:
          $ref: '#/components/responses/AIError'
  /api/v2/agent/conversations/{id}/messages:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
    get:
      operationId: listAgentMessages
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - items
                - next_before
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentMessage'
                  next_before:
                    type: integer
                    minimum: 0
        default:
          $ref: '#/components/responses/AIError'
      parameters:
      - name: before
        in: query
        schema:
          type: integer
          minimum: 0
    post:
      operationId: submitAgentQuestion
      responses:
        '202':
          description: Successful response.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - run_id
                - run
                properties:
                  run_id:
                    type: string
                  run:
                    $ref: '#/components/schemas/AgentRun'
        default:
          $ref: '#/components/responses/AIError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAgentQuestion'
  /api/v2/agent/runs/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
    get:
      operationId: getAgentRun
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - run
                - steps
                properties:
                  run:
                    $ref: '#/components/schemas/AgentRun'
                  steps:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentStep'
        default:
          $ref: '#/components/responses/AIError'
  /api/v2/agent/runs/{id}/cancel:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
    post:
      operationId: cancelAgentRun
      responses:
        '204':
          description: Completed; no response body.
        default:
          $ref: '#/components/responses/AIError'
  /api/v2/agent/subscription-drafts/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
    get:
      operationId: getAgentDraft
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionDraft'
        default:
          $ref: '#/components/responses/AIError'
    patch:
      operationId: editAgentDraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [version, payload]
              properties:
                version: {type: integer, minimum: 1}
                payload: {$ref: '#/components/schemas/AgentDraftInput'}
      responses:
        '200':
          description: New draft version; the previous version cannot be confirmed.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SubscriptionDraft'}
        default: {$ref: '#/components/responses/AIError'}
  /api/v2/agent/subscription-drafts/{id}/confirm:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
    post:
      operationId: confirmAgentDraft
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        default:
          $ref: '#/components/responses/AIError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - version
              properties:
                version:
                  type: integer
                  minimum: 1
  /api/v2/papers/{id}/ai-summary:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: integer, format: int64, minimum: 1}
    get:
      tags: [Papers]
      operationId: getPaperAISummary
      description: Reads current-version cached summaries after user paper authorization. Never calls a model.
      parameters:
        - name: language
          in: query
          schema: {type: string, enum: [zh, en]}
      responses:
        "200":
          description: Current summary state for each requested language.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AISummaryResponse"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {description: Operator accounts cannot access user AI endpoints.}
        "404": {description: Paper is not visible to the current user.}
        "503": {description: AI_DISABLED or AI_UNAVAILABLE; original paper remains available.}
    post:
      tags: [Papers]
      operationId: requestPaperAISummary
      description: Creates an on-demand user-scoped generation request using the current user's saved provider configuration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                language: {type: string, enum: [zh, en]}
      responses:
        "200":
          description: Cached result or completed failed attempt; no new task required.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AISummaryResponse"}
        "202":
          description: Generation is queued or processing.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AISummaryResponse"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {description: Operator accounts cannot access user AI endpoints.}
        "404": {description: Paper is not visible to the current user.}
        "503": {description: AI_DISABLED or AI_UNAVAILABLE.}
  /api/v2/ai/providers:
    get:
      tags: [Profile]
      operationId: listAIProviders
      responses:
        default: {$ref: "#/components/responses/AIError"}
        "200":
          description: Enabled fixed-endpoint providers and allowlisted models.
          content: {application/json: {schema: {$ref: "#/components/schemas/AIProviderPage"}}}
  /api/v2/ai/configuration:
    get:
      tags: [Profile]
      operationId: getAIConfiguration
      responses:
        default: {$ref: "#/components/responses/AIError"}
        "200":
          description: Masked current configuration.
          headers: {ETag: {$ref: "#/components/headers/AIConfigurationETag"}}
          content: {application/json: {schema: {$ref: "#/components/schemas/AIConfiguration"}}}
    put:
      tags: [Profile]
      operationId: putAIConfiguration
      description: Validates the submitted model once before atomically saving. An omitted or empty api_key reuses the existing secret only for the same provider. A new provider requires a key. Validation waits at most 30 seconds and never retries automatically.
      parameters: [{$ref: "#/components/parameters/OptionalAIIfMatch"}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/PutAIConfiguration"}}}}
      responses:
        default: {$ref: "#/components/responses/AIError"}
        "200":
          description: Tested active configuration.
          headers: {ETag: {$ref: "#/components/headers/AIConfigurationETag"}}
          content: {application/json: {schema: {$ref: "#/components/schemas/AIConfiguration"}}}
    patch:
      tags: [Profile]
      operationId: changeAIModel
      parameters: [{$ref: "#/components/parameters/AIIfMatch"}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [model], properties: {model: {type: string}}}}}}
      responses:
        default: {$ref: "#/components/responses/AIError"}
        "200":
          description: Tested active configuration.
          headers: {ETag: {$ref: "#/components/headers/AIConfigurationETag"}}
          content: {application/json: {schema: {$ref: "#/components/schemas/AIConfiguration"}}}
    delete:
      tags: [Profile]
      operationId: deleteAIConfiguration
      parameters: [{$ref: "#/components/parameters/AIIfMatch"}]
      responses:
        default: {$ref: "#/components/responses/AIError"}
        "204": {description: Configuration deleted and all user AI switches disabled.}
  /api/v2/ai/configuration/secret:
    put:
      tags: [Profile]
      operationId: rotateAISecret
      parameters: [{$ref: "#/components/parameters/AIIfMatch"}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [api_key], properties: {api_key: {type: string, writeOnly: true, maxLength: 1024}}}}}}
      responses:
        default: {$ref: "#/components/responses/AIError"}
        "200":
          description: Tested active configuration.
          headers: {ETag: {$ref: "#/components/headers/AIConfigurationETag"}}
          content: {application/json: {schema: {$ref: "#/components/schemas/AIConfiguration"}}}
  /api/v2/ai/configuration/test:
    post:
      tags: [Profile]
      operationId: testAIConfiguration
      responses:
        default: {$ref: "#/components/responses/AIError"}
        "200":
          description: Connection test passed.
          headers: {ETag: {$ref: "#/components/headers/AIConfigurationETag"}}
          content: {application/json: {schema: {$ref: "#/components/schemas/AIConfiguration"}}}
        "409": {$ref: "#/components/responses/AIError"}
        "429": {$ref: "#/components/responses/AIError"}
  /api/v2/ai/usage:
    get:
      tags: [Profile]
      operationId: getAIUsage
      parameters:
        - {name: from, in: query, schema: {type: string, format: date}}
        - {name: to, in: query, schema: {type: string, format: date}}
      responses:
        default: {$ref: "#/components/responses/AIError"}
        "200": {description: Up to 30 UTC days of per-feature usage., content: {application/json: {schema: {$ref: "#/components/schemas/AIUsagePage"}}}}
  /api/v2/ai/calls:
    get:
      tags: [Profile]
      operationId: listAICalls
      parameters:
        - {name: page, in: query, schema: {type: integer, minimum: 1, maximum: 100000, default: 1}}
        - {name: page_size, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
        - {name: feature, in: query, schema: {type: string, enum: [config_test, paper, digest, subscription_agent, paper_qa]}}
        - {name: status, in: query, schema: {type: string, enum: [started, succeeded, failed, unknown]}}
      responses:
        default: {$ref: "#/components/responses/AIError"}
        "200":
          description: User-isolated call diagnostics. Detailed records are retained for 30 days.
          content: {application/json: {schema: {$ref: "#/components/schemas/AICallPage"}}}
...
  /api/v2/auth/register:
    post:
      tags: [Authentication]
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/EmailAlreadyRegistered"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v2/auth/login:
    post:
      tags: [Authentication]
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Credentials accepted. Sets a host-only HttpOnly SameSite=Lax signalwatch_session cookie scoped to /api/v2/auth, Secure in production, with a fixed default seven-day expiry. The secret is never returned in JSON. Responses use Cache-Control no-store.
          headers:
            Set-Cookie:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v2/sources:
    get:
      tags: [Sources]
      operationId: listSources
      responses:
        "200":
          description: Enabled public source catalog entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicSource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: signalwatch_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 access token returned by the login endpoint.
...
  schemas:
...
    AIProviderPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/AIProvider"}
...
  responses:
    AIError:
      description: Structured AI error. 409 denotes conflict or missing configuration; 422 invalid credentials/model; 429 local daily limit, admission interval or provider rate limit; 502 provider failure; 504 provider timeout; 503 application/storage unavailable. Retry-After is returned when a retry time is known. No model call is retried automatically.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
...